+++ packages/traceloop-sdk/traceloop/sdk/tracing/tracing.py
@@ -390,6 +390,7 @@ def init_instrumentations():
     init_anthropic_instrumentor()
     init_cohere_instrumentor()
     init_pinecone_instrumentor()
+    init_qdrant_instrumentor()
     init_chroma_instrumentor()
     init_haystack_instrumentor()
     init_langchain_instrumentor()
@@ -446,6 +447,16 @@ def init_pinecone_instrumentor():
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
     return True
+
+
+def init_qdrant_instrumentor():
+    if importlib.util.find_spec("qdrant_client") is not None:
+        Telemetry().capture("instrumentation:qdrant:init")
+        from opentelemetry.instrumentation.qdrant import QdrantInstrumentor
+
+        instrumentor = QdrantInstrumentor()
+        if not instrumentor.is_instrumented_by_opentelemetry:
+            instrumentor.instrument()
 
 
 def init_chroma_instrumentor():